Refactor into a library and a binary
