+++ mbt-utils/src/lib.rs
@@ -1,7 +1,7 @@
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        assert_eq!(2 + 2, 4);
-    }
-}
+pub mod commit;
+pub mod consensus;
+pub mod header;
+pub mod helpers;
+pub mod producer;
+pub mod validator;
+
